fix:update manifest.xml version correctly
The manifest.xml file was not being updated during the release process
because the XPath expression in the release-please-config.json file did
not account for the default namespace in the manifest.xml file.

This change updates the XPath to be namespace-agnostic by using the
`local-name()` function. This ensures that the version element is
correctly located and updated during the release process.

Also fixed a few release please warnings.